fix: separate interfaces from statistic model
