+++ rslang/src/app/statistics/components/statistics-page/statistics-page.component.html
@@ -1,9 +1,14 @@
 <div class="statistic-page">
   <h1>Statistic</h1>
   <h2>Short-term statistic</h2>
-  <app-short-term-statistic
-    *ngFor="let statistic of statisticsList"
-    [statistic]="statistic"
-  ></app-short-term-statistic>
+  <mat-tab-group mat-align-tabs="center">
+    <mat-tab *ngFor="let statistic of statisticsList; let index = index" [label]=statistic.date.toDateString()>
+      <app-short-term-statistic
+        [statistic]="statistic"
+      ></app-short-term-statistic>
+    </mat-tab>
+  </mat-tab-group>
+  <h2>Long-term statistic</h2>
   <app-long-term-statistic></app-long-term-statistic>
 </div>
+
