+++ rslang/src/app/statistics/models/game-statistic.model.ts
@@ -1,0 +1,5 @@
+export interface GameStatistic {
+  name: string;
+  winrate: number;
+  streak: number;
+}
