+++ rslang/src/app/statistics/models/statistic.model.ts
@@ -1,17 +1,8 @@
+import { GameStatistic } from './game-statistic.model';
+import { WordStatistic } from './word-statistic.model';
+
 export interface Statistic {
   date: Date;
   games: Array<GameStatistic>;
   learnedWords: Array<WordStatistic>;
-}
-
-export interface GameStatistic {
-  name: string;
-  winrate: number;
-  streak: number;
-}
-
-interface WordStatistic {
-  id: string;
-  group: number;
-  page: number;
 }
