+++ rslang/src/app/statistics/models/word-statistic.model.ts
@@ -1,0 +1,5 @@
+export interface WordStatistic {
+  id: string;
+  group: number;
+  page: number;
+}
